Make batch version into the main RnaChipIntegrator utility (i.e. only batch version now exists).
