+++ rnachipintegrator/cli.py
@@ -33,11 +33,13 @@ import analysis
 import output
 import xls_output
 import logging
-logging.getLogger().setLevel(logging.WARNING)
-logging.basicConfig(format='%(levelname)s: %(message)s')
+from multiprocessing import Pool
 
 from . import get_version
 __version__ = get_version()
+
+logging.getLogger().setLevel(logging.WARNING)
+logging.basicConfig(format='%(levelname)s: %(message)s')
 
 #######################################################################
 # Data
@@ -647,6 +649,18 @@ def find_nearest_peaks(params):
             params.only_differentially_expressed):
         yield (gene,peaks,params)
 
+def find_nearest_features_as_list(params):
+    """
+    Wrapper to fetch results from 'find_nearest_features' as a list
+    """
+    return list(find_nearest_features(params))
+
+def find_nearest_peaks_as_list(params):
+    """
+    Wrapper to fetch results from 'find_nearest_peaks' as a list
+    """
+    return list(find_nearest_peaks(params))
+
 #######################################################################
 # Main program
 #######################################################################
@@ -660,31 +674,30 @@ def main(args=None):
     if args is None:
         args = sys.argv[1:]
 
-    p = CLI(usage="%(prog)s [options] GENES PEAKS",
+    p = CLI(usage="%(prog)s [options] [GENES PEAKS]",
             description=
             "Analyse GENES (any set of genes or genomic "
-            "features) against one or more sets of PEAKS "
-            "(a set of regions) and report nearest genes "
-            "to each peak (and vice versa)")
+            "features) against PEAKS (a set of regions) "
+            "and report nearest genes to each peak (and "
+            "vice versa)")
 
-    # Analysis options
     p.add_option_group("Analysis options")
-    p.add_option('--cutoff',action='store',dest='max_distance',
-                 type=int,default=_DEFAULTS.CUTOFF,
-                 help="Maximum distance allowed between peaks "
-                 "and genes before being omitted from the "
-                 "analyses (default %dbp; set to zero for no "
-                 "cutoff)" % _DEFAULTS.CUTOFF,
+    p.add_option('--cutoffs',action='store',dest='cutoffs',
+                 default=None,
+                 help="Comma-separated list of one or more "
+                 "maximum distances allowed between peaks "
+                 "and genes (bp). An analysis will be "
+                 "performed for each GENES-PEAKS pair at "
+                 "each cutoff distance (default %dbp; set "
+                 "to zero for no cutoff)" % _DEFAULTS.CUTOFF,
                  group="Analysis options")
     p.add_edge_option(group="Analysis options")
     p.add_only_de_option(group="Analysis options")
 
-    # Reporting options
     p.add_option_group("Reporting options")
     p.add_number_option(group="Reporting options")
     p.add_promoter_region_option(group="Reporting options")
 
-    # Output options
     p.add_option_group("Output options")
     p.add_name_option(group="Output options")
     p.add_compact_option(group="Output options")
@@ -692,7 +705,25 @@ def main(args=None):
     p.add_pad_option(group="Output options")
     p.add_xlsx_option(group="Output options")
 
-    # Advanced options
+    p.add_option_group("Batch options")
+    p.add_option('--genes',action='store',dest="genes",nargs="+",
+                 metavar="GENES_FILE",
+                 help="Specify multiple genes files (if used then "
+                 "peaks file(s) must be specified using --peaks "
+                 "option)",
+                 group="Batch options")
+    p.add_option('--peaks',action='store',dest="peaks",nargs="+",
+                 metavar="PEAKS_FILE",
+                 help="Specify multiple peaks files (if used then "
+                 "genes file(s) must be specified using --genes "
+                 "option)",
+                 group="Batch options")
+    p.add_option('-n','--nprocessors',action='store',
+                 type=int,dest='nprocs',default=1,
+                 help="Number of processors/cores to run the "
+                 "program using (default: 1)",
+                 group="Batch options")
+
     p.add_option_group("Advanced options")
     p.add_analyses_option(group="Advanced options")
     p.add_feature_option(group="Advanced options")
@@ -703,23 +734,48 @@ def main(args=None):
     options,args = p.parse_args()
 
     # Input files
-    if len(args) != 2:
-        p.error("need to supply 2 files (genes and peaks)")
-    gene_file,peak_file = args
+    peak_files = []
+    gene_files = []
+    if options.peaks:
+        peak_files = [f for f in options.peaks]
+    if options.genes:
+        gene_files = [f for f in options.genes]
+    if peak_files and gene_files:
+        if len(args) > 0:
+            p.error("too many arguments: files already supplied via "
+                    "--peaks and --genes")
+    elif peak_files:
+        p.error("must supply GENES file(s) via --genes when using "
+                "--peaks option")
+    elif gene_files:
+        p.error("must supply PEAKS file(s) via --peaks when using "
+                "--genes option")
+    elif len(args) == 2:
+        gene_files.append(args[0])
+        peak_files.append(args[1])
+    else:
+        p.error("need to supply genes and peaks via command line "
+                "or via --genes and --peaks options")
 
     # Report version and authors
     print p.get_version()
     print _PROGRAM_INFO
 
-    # Promoter region
-    promoter = (abs(int(options.promoter_region.split(',')[0])),
-                abs(int(options.promoter_region.split(',')[1])))
+    # Process cutoffs
+    if options.cutoffs is None:
+        cutoffs = [_DEFAULTS.CUTOFF,]
+    else:
+        cutoffs = list()
+        for cutoff in str(options.cutoffs).split(','):
+            try:
+                cutoffs.append(int(cutoff))
+            except ValueError:
+                logging.critical("Bad cutoff value: '%s'" % cutoff)
+                sys.exit(1)
+    cutoffs.sort()
 
-    # Reporting options
-    max_distance = options.max_distance
-    if max_distance <= 0:
-        max_distance = None
-    max_closest = options.max_closest
+    # Deal with zero cutoff distance meaning 'no cutoff'
+    cutoffs = [d if d != 0 else None for d in cutoffs]
 
     # Gene edge to use
     if options.edge == 'tss':
@@ -727,13 +783,11 @@ def main(args=None):
     else:
         tss_only = False
 
-    # Feature type
-    if options.feature_type is None:
-        feature_type = 'gene'
-    else:
-        feature_type = options.feature_type
+    # Promoter region
+    promoter = (abs(int(options.promoter_region.split(',')[0])),
+                abs(int(options.promoter_region.split(',')[1])))
 
-    # Columns to extract from input peaks file
+    # Columns for peak data
     if options.peak_cols is None:
         peak_cols = (1,2,3)
     else:
@@ -741,17 +795,19 @@ def main(args=None):
             peak_cols = tuple([int(x)
                                for x in options.peak_cols.split(',')])
         except Exception as ex:
-            p.error("Bad column assignment for --peak_cols")
+            logging.fatal("Bad column assignment for --peak_cols")
+            sys.exit(1)
 
-    # Handle peak IDs, if specified
-    if options.peak_id is not None:
+    # Use IDs for peaks?
+    try:
         peak_id_col = int(options.peak_id)
-    else:
+    except TypeError:
         peak_id_col = None
 
-    # Reporting formats
+    # Set up reporting
     if options.compact:
         mode = output.SINGLE_LINE
+        placeholder = '.'
         if peak_id_col is None:
             peak_fields = ('peak.chr','peak.start','peak.end',
                            'list(feature.id,strand,TSS,TES,dist_closest,'
@@ -770,13 +826,13 @@ def main(args=None):
                            'feature.end','feature.strand',
                            'list(peak.id,chr,start,end,dist_closest,dist_TSS,'
                            'direction,in_the_feature)')
-        placeholder = '.'
         if options.summary:
             options.summary = False
             logging.error("--summary option not compatible with --compact")
             sys.exit(1)
     else:
         mode = output.MULTI_LINE
+        placeholder = '---'
         if peak_id_col is None:
             peak_fields = ('peak.chr','peak.start','peak.end',
                            'feature.id','strand','TSS','TES',
@@ -793,111 +849,176 @@ def main(args=None):
                            'direction','overlap_feature','overlap_promoter')
             gene_fields = ('feature.id','feature.chr','feature.start',
                            'feature.end','feature.strand',
-                           'peak.id','chr','start','end','dist_closest','dist_TSS',
-                           'direction','in_the_feature')
-        placeholder = '---'
+                           'peak.id','chr','start','end',
+                           'dist_closest','dist_TSS','direction',
+                           'in_the_feature')
+
+    # Update fields for batch mode
+    if len(cutoffs) > 1:
+        peak_fields = tuple(['cutoff']+list(peak_fields))
+        gene_fields = tuple(['cutoff']+list(gene_fields))
+    if len(gene_files) > 1:
+        peak_fields = tuple(['feature_file']+list(peak_fields))
+        gene_fields = tuple(['feature_file']+list(gene_fields))
+    if len(peak_files) > 1:
+        peak_fields = tuple(['peak_file']+list(peak_fields))
+        gene_fields = tuple(['peak_file']+list(gene_fields))
 
     # Analyses to run
     peak_centric = (options.analyses in ("all","peak_centric",))
     gene_centric = (options.analyses in ("all","gene_centric",))
 
-    # Report settings
-    print "Input genes file: %s" % gene_file
-    print "Input peaks file: %s" % peak_file
-    print
-    if max_distance is not None:
-        print "Maximum cutoff distance: %d (bp)" % max_distance
-    else:
-        print "Maximum cutoff distance: no cutoff"
-    print "Maximum no. of hits    : %s" % ('All' if max_closest is None
-                                           else "%d" % max_closest)
-    print "Promoter region        : -%d to %d (bp from TSS)" % promoter
+    # Report inputs
+    print "Genes files    : %s" % gene_files[0]
+    for gene_file in gene_files[1:]:
+        print "                 %s" % gene_file
+    print "Peaks files    : %s" % peak_files[0]
+    for peak_file in peak_files[1:]:
+        print "                 %s" % peak_file
+    print "Cutoffs (bp)   : %s" % ','.join([str(d) if d != 0
+                                            else "no cutoff"
+                                            for d in cutoffs])
+    print "Edge           : %s" % ('TSS only' if tss_only
+                                   else 'TSS or TES')
+    print "DE only        : %s" % ('yes' if options.only_diff_expressed
+                                   else 'no')
+    print "Nprocs         : %s" % options.nprocs
+    print "Max no. of hits: %s" % ('All' if options.max_closest is None
+                                   else "%d" % options.max_closest)
+    print "Promoter region: -%d to %d (bp from TSS)" % promoter
+    print "Feature type   : %s" % options.feature_type
     print
     print "Analyses:"
     print "- Peak-centric: %s" % ('yes' if peak_centric else 'no')
     print "- Gene-centric: %s" % ('yes' if gene_centric else 'no')
-    if tss_only:
-        print "Distances will be calculated from gene TSS only"
-    else:
-        print "Distances will be calculated from nearest of gene TSS or TES"
-    print
-    print "Genomic features are '%ss'" % feature_type
-    print
 
     # Read in gene data
-    genes = read_feature_file(gene_file)
-    if options.only_diff_expressed:
-        if not genes.isFlagged():
+    gene_lists = dict()
+    for gene_file in gene_files:
+        genes = read_feature_file(gene_file)
+        if options.only_diff_expressed and not genes.isFlagged():
             logging.fatal("--only-DE flag needs input genes flagged as "
                           "differentially expressed")
             sys.exit(1)
-        else:
-            print
-            print "*** Only differentially expressed genes will used ***"
-    print
+        gene_lists[gene_file] = genes
 
     # Read in peak data
-    peaks = read_peak_file(peak_file,
-                           peak_cols=peak_cols,
-                           peak_id_col=peak_id_col)
+    peak_lists = dict()
+    for peak_file in peak_files:
+        peak_lists[peak_file] = read_peak_file(peak_file,
+                                               peak_cols=peak_cols,
+                                               peak_id_col=peak_id_col)
 
-    # Set up output files
+    # Output files
     if options.name is not None:
         basename = options.name
     else:
-        basename = os.path.splitext(os.path.basename(gene_file))[0]
+        basename = os.path.splitext(os.path.basename(gene_files[0]))[0]
     outputs = OutputFiles(basename)
-
-    # Clean up any pre-existing output files that would otherwise
-    # be overwritten
     outputs.remove_files()
 
-    # Do the analyses
+    # Assemble inputs over cutoffs and peak and gene lists
+    # The same parameters can be used for both peak- and
+    # gene-centric analyses
+    analysis_params = []
+    for peaks in peak_files:
+        for genes in gene_files:
+            for cutoff in cutoffs:
+                analysis_params.append(
+                    AnalysisParams(
+                        genes=gene_lists[genes],
+                        peaks=peak_lists[peaks],
+                        cutoff=cutoff,
+                        tss_only=tss_only,
+                        only_differentially_expressed=
+                        options.only_diff_expressed
+                    ))
+
+    # Run the analyses
     if peak_centric:
         print "**** Peak-centric analysis: nearest genes to each peak ****"
+        # Build reporter
         reporter = output.AnalysisReportWriter(
             mode,peak_fields,
             promoter_region=promoter,
             null_placeholder=placeholder,
-            max_hits=max_closest,
+            max_hits=options.max_closest,
             pad=options.pad_output,
             outfile=outputs.peak_centric_out,
             summary=(outputs.peak_centric_summary
                      if options.summary else None),
-            feature_type=feature_type)
-        params = AnalysisParams(genes=genes,
-                                peaks=peaks,
-                                cutoff=max_distance,
-                                tss_only=tss_only,
-                                only_differentially_expressed=
-                                options.only_diff_expressed)
-        for peak,nearest_genes,prms in find_nearest_features(params):
-            reporter.write_nearest_features(peak,nearest_genes)
+            feature_type=options.feature_type)
+        # Run the analyses
+        if options.nprocs > 1:
+            # Multiple cores
+            pool = Pool(options.nprocs)
+            # Version of multiprocessing which can also
+            # handle ctrl-C terminating the program
+            # See http://bryceboe.com/2010/08/26/python-multiprocessing-and-keyboardinterrupt/
+            p = pool.map_async(find_nearest_features_as_list,
+                               analysis_params)
+            try:
+                results = p.get(0xFFFF)
+            except KeyboardInterrupt:
+                print "KeyboardInterrupt"
+                sys.exit(1)
+        else:
+            # Single core
+            results = map(lambda p: list(find_nearest_features(p)),
+                          analysis_params)
+        # Output the results
+        for result in results:
+            for peak,nearest_genes,params in result:
+                reporter.write_nearest_features(
+                    peak,nearest_genes,
+                    peak_file=params.peaks.source_file,
+                    feature_file=params.genes.source_file,
+                    cutoff=params.cutoff)
         reporter.close()
         print "Results written to %s" % outputs.peak_centric_out
         if options.summary:
-            print "Summary written to %s" % peak_centric_summary
+            print "Summary written to %s" % outputs.peak_centric_summary
         print
 
+    # Run the analyses
     if gene_centric:
         print "**** Gene-centric analysis: nearest peaks to each gene ****"
+        # Build reporter
         reporter = output.AnalysisReportWriter(
             mode,gene_fields,
             null_placeholder=placeholder,
-            max_hits=max_closest,
+            max_hits=options.max_closest,
             pad=options.pad_output,
             outfile=outputs.gene_centric_out,
             summary=(outputs.gene_centric_summary
                      if options.summary else None),
-            feature_type=feature_type)
-        params = AnalysisParams(genes=genes,
-                                peaks=peaks,
-                                cutoff=max_distance,
-                                tss_only=tss_only,
-                                only_differentially_expressed=
-                                options.only_diff_expressed)
-        for gene,nearest_peaks,prms in find_nearest_peaks(params):
-            reporter.write_nearest_peaks(gene,nearest_peaks)
+            feature_type=options.feature_type)
+        # Run the analyses
+        if options.nprocs > 1:
+            # Multiple cores
+            pool = Pool(options.nprocs)
+            # Version of multiprocessing which can also
+            # handle ctrl-C terminating the program
+            # See http://bryceboe.com/2010/08/26/python-multiprocessing-and-keyboardinterrupt/
+            p = pool.map_async(find_nearest_peaks_as_list,
+                               analysis_params)
+            try:
+                results = p.get(0xFFFF)
+            except KeyboardInterrupt:
+                print "KeyboardInterrupt"
+                sys.exit(1)
+        else:
+            # Single core
+            results = map(lambda p: list(find_nearest_peaks(p)),
+                          analysis_params)
+        # Output the results
+        for result in results:
+            for gene,nearest_peaks,params in result:
+                reporter.write_nearest_features(
+                    gene,nearest_peaks,
+                    peak_file=params.peaks.source_file,
+                    feature_file=params.genes.source_file,
+                    cutoff=params.cutoff)
         reporter.close()
         print "Results written to %s" % outputs.gene_centric_out
         if options.summary:
@@ -909,19 +1030,22 @@ def main(args=None):
         print "**** Writing XLSX file ****"
         xlsx = xls_output.XLSX(outputs.xlsx_out,
                                p.get_version(),
-                               feature_type)
+                               options.feature_type)
         # Write the settings
-        xlsx.append_to_notes("Input %ss file\t%s" % (feature_type,
-                                                     gene_file))
-        xlsx.append_to_notes("Input peaks file\t%s" % peak_file)
-        if max_distance is not None:
-            xlsx.append_to_notes("Maximum cutoff distance (bp)\t%d" %
-                                 max_distance)
-        else:
-            xlsx.append_to_notes("Maximum cutoff distance (bp)\tno cutoff")
+        xlsx.append_to_notes("Input %ss file\t%s" % (options.feature_type,
+                                                     gene_files[0]))
+        for gene_file in gene_files[1:]:
+            xlsx.append_to_notes("\t%s" % gene_file)
+        xlsx.append_to_notes("Input peaks file\t%s" % peak_files[0])
+        for peak_file in peak_files[1:]:
+            xlsx.append_to_notes("\t%s" % peak_file)
+        cutoff_distances = [str(d) if d != 0 else 'no cutoff'
+                            for d in cutoffs]
+        xlsx.append_to_notes("Cutoff distances (bp)\t%s" %
+                             ','.join([str(d) for d in cutoff_distances]))
         xlsx.append_to_notes("Maximum no. of hits to report\t%s"
-                             % ('All' if max_closest is None
-                                else "%d" % max_closest))
+                             % ('All' if options.max_closest is None
+                                else "%d" % options.max_closest))
         xlsx.append_to_notes("Promoter region (bp from TSS)\t-%d to %d" %
                              promoter)
         if tss_only:
@@ -929,8 +1053,9 @@ def main(args=None):
         else:
             xlsx.append_to_notes("Distances calculated from\tTSS or TES")
         xlsx.append_to_notes("Only use differentially expressed %ss\t%s" %
-                             (feature_type,
-                              "Yes" if options.only_diff_expressed else "No"))
+                             (options.feature_type,
+                              "Yes" if options.only_diff_expressed
+                              else "No"))
         # Add features to peaks
         if peak_centric:
             xlsx.write_peak_centric(peak_fields)
@@ -938,23 +1063,24 @@ def main(args=None):
             if options.summary:
                 xlsx.append_to_notes("\n'Peak-centric (summary)' lists the "
                                      "'top' result (i.e. closest peak/%s "
-                                     "pair) for each peak" % feature_type)
+                                     "pair) for each peak" %
+                                     options.feature_type)
                 xlsx.add_result_sheet('Peak-centric (summary)',
-                                      peak_centric_summary)
+                                      outputs.peak_centric_summary)
         # Add peaks to features
         if gene_centric:
             xlsx.write_feature_centric(gene_fields)
-            xlsx.add_result_sheet('%s-centric' % feature_type.title(),
+            xlsx.add_result_sheet('%s-centric' % options.feature_type.title(),
                                   outputs.gene_centric_out)
             if options.summary:
                 xlsx.append_to_notes("\n'%s-centric (summary)' lists the "
                                      "'top' result (i.e. closest %s/peak "
                                      "pair) for each %s" %
-                                     (feature_type.title(),
-                                      feature_type,
-                                      feature_type))
+                                     (options.feature_type.title(),
+                                      options.feature_type,
+                                      options.feature_type))
                 xlsx.add_result_sheet('%s-centric (summary)' %
-                                      feature_type.title(),
+                                      options.feature_type.title(),
                                       outputs.gene_centric_summary)
         xlsx.write()
         print "Wrote %s" % outputs.xlsx_out
