+++ setup.py
@@ -35,8 +35,7 @@ setup(
     maintainer_email = 'peter.briggs@manchester.ac.uk',
     packages = ['rnachipintegrator'],
     entry_points = { 'console_scripts': [
-        'RnaChipIntegrator = rnachipintegrator.cli:main',
-        'RnaChipIntegrator-batch = rnachipintegrator.batch:main']
+        'RnaChipIntegrator = rnachipintegrator.cli:main',]
     },
     license = 'Artistic License',
     keywords = ['RnaChipIntegrator',],
